not important

+++ src/loading/data_loader/data_loader.py
@@ -77,5 +77,5 @@ class DataLoader(TorchDataLoader):
         return {
             "image": torch.stack(
                 [e["image"] for e in data]
-            ).to(memory_format=torch.contiguous_format).type(torch.float16)
+            ).to(memory_format=torch.contiguous_format),  # .type(torch.float16)
         }
